+++ src/chaosdetection/lyapunovs.jl
@@ -30,7 +30,8 @@ See also [`lyapunov`](@ref), [`local_growth_rates`](@ref).
 * `Δt = 1` : Time of individual evolutions
   between successive orthonormalization steps. For continuous systems this is approximate.
 * `show_progress = false` : Display a progress bar of the process.
-* `diffeq...` : Keyword arguments propagated into `init` of DifferentialEquations.jl.
+* `diffeq` is a `NamedTuple` (or `Dict`) of keyword arguments propagated into
+  `init` of DifferentialEquations.jl.
   See [`trajectory`](@ref) for examples. Only valid for continuous systems.
 
 ## Description
@@ -70,14 +71,20 @@ lyapunovspectrum(ds::DS, N, k::Int = dimension(ds); kwargs...) =
 lyapunovspectrum(ds, N, orthonormal(dimension(ds), k); kwargs...)
 
 function lyapunovspectrum(ds::DS{IIP, S, D}, N, Q0::AbstractMatrix; 
-        Ttr::Real = 0, Δt::Real = 1, u0 = get_state(ds), show_progress = false, diffeq...
+        Ttr::Real = 0, Δt::Real = 1, u0 = get_state(ds), show_progress = false, 
+        diffeq = NamedTuple(), kwargs...
     ) where {IIP, S, D}
+
+    if !isempty(kwargs)
+        @warn DIFFEQ_DEP_WARN
+        diffeq = NamedTuple(kwargs)
+    end
 
     if typeof(ds) <: DDS
         @assert typeof(Ttr) == Int
         integ = tangent_integrator(ds, Q0; u0)
     else
-        integ = tangent_integrator(ds, Q0; u0, diffeq...)
+        integ = tangent_integrator(ds, Q0; u0, diffeq)
     end
     λ = lyapunovspectrum(integ, N, Δt, Ttr, show_progress)
     return λ
@@ -145,34 +152,103 @@ function lyapunovspectrum(ds::DDS{false, T, 1}, N; Ttr = 0) where {T}
     return λ/N
 end
 
-function lyapunovspectrum_convergence(integ, N, dt::Real, Ttr::Real = 0.0)
+#####################################################################################
+#                           Lyapunov Spectrum Convergence                           #
+#####################################################################################
+"""
+    lyapunovspectrum_convergence(ds::DynamicalSystem, N [, k::Int | Q0]; kwargs...) -> λs, t
 
-    T = stateeltype(integ)
-    t0 = integ.t
+Computes the convergence of the Lyapunov exponents, calculated exactly as in `lyapunovspectrum`.
+(See [`lyapunovspectrum`](@ref) for information on the parameters and implementation).
+Returns `λs`, a vector of `N` vectors, wherein vector of index `i` contains the Lyapunov
+spectrum evaluated up to the time `t[i]`. The final vector is equal to the returned vector
+of `lyapunovspectrum`.
+
+## Performance Notes
+This function uses a [`tangent_integrator`](@ref).
+For loops over initial conditions and/or
+parameter values one should use the low level method that accepts
+an integrator, and `reinit!` it to new initial conditions.
+See the "advanced documentation" for info on the integrator object.
+The low level method is
+```julia
+lyapunovspectrum_convergence(tinteg, N, Δt::Real, Ttr::Real)
+```
+"""
+lyapunovspectrum_convergence(ds::DS, N, k::Int = dimension(ds); kwargs...) =
+lyapunovspectrum_convergence(ds, N, orthonormal(dimension(ds), k); kwargs...)
+
+function lyapunovspectrum_convergence(ds::DS{IIP, S, D}, N, Q0::AbstractMatrix;
+        Ttr::Real = 0, Δt::Real = 1, u0 = get_state(ds), show_progress = false, diffeq...
+    ) where {IIP, S, D}
+
+    if typeof(ds) <: DDS
+        @assert typeof(Ttr) == Int
+        integ = tangent_integrator(ds, Q0; u0)
+    else
+        integ = tangent_integrator(ds, Q0; u0, diffeq...)
+    end
+    λ = lyapunovspectrum_convergence(integ, N, Δt, Ttr, show_progress)
+    return λ
+end
+
+function lyapunovspectrum_convergence(integ, N, Δt::Real, Ttr::Real = 0.0, show_progress=false)
+    if show_progress
+        progress = ProgressMeter.Progress(N; desc = "Lyapunov Spectrum: ", dt = 1.0)
+    end
+    B = copy(get_deviations(integ)) # for use in buffer
     if Ttr > 0
+        t0 = integ.t
         while integ.t < t0 + Ttr
-            step!(integ, dt)
-            qrdec = LinearAlgebra.qr(get_deviations(integ))
-            set_deviations!(integ, qrdec.Q)
+            step!(integ, Δt)
+            Q, R = _buffered_qr(B, get_deviations(integ))
+            set_deviations!(integ, Q)
         end
     end
+
     k = size(get_deviations(integ))[2]
+    T = stateeltype(integ)
     t0 = integ.t; t = zeros(T, N); t[1] = t0
     λs = [zeros(T, k) for i in 1:N];
 
     for i in 2:N
-        step!(integ, dt)
-        qrdec = LinearAlgebra.qr(get_deviations(integ))
+        step!(integ, Δt)
+        Q, R = _buffered_qr(B, get_deviations(integ))
         for j in 1:k
-            @inbounds λs[i][j] = λs[i-1][j] + log(abs(qrdec.R[j,j]))
+            @inbounds λs[i][j] = λs[i-1][j] + log(abs(R[j,j]))
         end
         t[i] = integ.t
-        set_deviations!(integ, qrdec.Q)
+        set_deviations!(integ, Q)
+        show_progress && ProgressMeter.update!(progress, i)
     end
     popfirst!(λs); popfirst!(t)
     for j in eachindex(t)
         λs[j] ./= (t[j] - t0)
     end
+    return λs, t
+end
+
+lyapunovspectrum_convergence(ds::DiscreteDynamicalSystem{IIP, T, 1}, a...; kw...) where {IIP, T} =
+error("For discrete 1D systems, only method with state type = number is implemented.")
+
+function lyapunovspectrum_convergence(ds::DDS{false, T, 1}, N; Ttr = 0) where {T}
+    x = get_state(ds); f = ds.f
+    p = ds.p; t0 = ds.t0
+    if Ttr > 0
+        for i in t0:(Ttr+t0)
+            x = f(x, p, i)
+        end
+    end
+    t = (t0+Ttr):(t0+Ttr+N)
+    λs = zeros(T, length(t));
+    for idx = 2:length(t)
+        i = t[idx]
+        x = f(x, p, i)
+        @inbounds λs[idx] = λs[idx-1] + log(abs(ds.jacobian(x, p, i)))
+    end
+    t = collect(t)
+    popfirst!(λs); popfirst!(t)
+    λs ./= t .- t0
     return λs, t
 end
 
@@ -210,7 +286,8 @@ See also [`lyapunovspectrum`](@ref), [`local_growth_rates`](@ref).
   of the system). This function can be used when you want to avoid
   the test state appearing in a region of the phase-space where it would have
   e.g. different energy or escape to infinity.
-* `diffeq...` : Keyword arguments propagated into `init` of DifferentialEquations.jl.
+* `diffeq` is a `NamedTuple` (or `Dict`) of keyword arguments propagated into
+  `init` of DifferentialEquations.jl.
   See [`trajectory`](@ref) for examples. Only valid for continuous systems.
 
 
@@ -243,15 +320,20 @@ lyapunov(pinteg, T, Ttr, Δt, d0, ut, lt)
 [^Benettin1976]: G. Benettin *et al.*, Phys. Rev. A **14**, pp 2338 (1976)
 """
 function lyapunov(ds::DS, T;
-                  u0 = get_state(ds),
-                  Ttr = 0,
-                  d0=1e-9,
-                  upper_threshold = 1e-6,
-                  lower_threshold = 1e-12,
-                  inittest = inittest_default(dimension(ds)),
-                  Δt = 1,
-                  diffeq...
-                  )
+        u0 = get_state(ds),
+        Ttr = 0,
+        d0=1e-9,
+        upper_threshold = 1e-6,
+        lower_threshold = 1e-12,
+        inittest = inittest_default(dimension(ds)),
+        Δt = 1,
+        diffeq = NamedTuple(), kwargs...
+    )
+
+    if !isempty(kwargs)
+        @warn DIFFEQ_DEP_WARN
+        diffeq = NamedTuple(kwargs)
+    end
 
     ST = stateeltype(ds)
     lower_threshold ≤ d0 ≤ upper_threshold || throw(ArgumentError(
@@ -260,7 +342,7 @@ function lyapunov(ds::DS, T;
     if typeof(ds) <: DDS
         pinteg = parallel_integrator(ds, [deepcopy(u0), inittest(u0, d0)])
     else
-        pinteg = parallel_integrator(ds, [deepcopy(u0), inittest(u0, d0)]; diffeq...)
+        pinteg = parallel_integrator(ds, [deepcopy(u0), inittest(u0, d0)]; diffeq)
     end
     λ::ST = lyapunov(pinteg, T, Ttr, Δt, d0, upper_threshold, lower_threshold)
     return λ
@@ -376,18 +458,25 @@ The output of this function is sometimes referred as "Nonlinear Local Lyapunov E
   outputs a pertrubation vector (preferrably `SVector`) given the system, current initial
   condition `u` and the counter `j ∈ 1:S`. If not given, a random perturbation is
   generated with norm given by the keyword `e = 1e-6`.
-* `diffeq...`: Keywords propagated to the solvers of DifferentialEquations.jl.
+* `diffeq` is a `NamedTuple` (or `Dict`) of keyword arguments propagated into
+  `init` of DifferentialEquations.jl.
+  See [`trajectory`](@ref) for examples. Only valid for continuous systems.
 """
 function local_growth_rates(ds::DynamicalSystem, points;
         S = 100, Δt = 5, e = 1e-6,
         perturbation = (ds, u, j) -> _random_Q0(ds, u, j, e),
-        diffeq...
+        diffeq = NamedTuple(), kwargs...
     )
+
+    if !isempty(kwargs)
+        @warn DIFFEQ_DEP_WARN
+        diffeq = NamedTuple(kwargs)
+    end
+
     λlocal = zeros(length(points), S)
-    D = dimension(ds)
     Q0 = perturbation(ds, points[1], 1)
     states = [points[1], points[1] .+ Q0]
-    pinteg = parallel_integrator(ds, states; diffeq...)
+    pinteg = parallel_integrator(ds, states; diffeq)
 
     for (i, u) in enumerate(points)
         for j in 1:S
@@ -406,7 +495,7 @@ end
 
 function _random_Q0(ds, u, j, e)
     D, T = dimension(ds), eltype(ds)
-    Q0 = randn(Random.GLOBAL_RNG, SVector{D, eltype(ds)})
+    Q0 = randn(Random.GLOBAL_RNG, SVector{D, T})
     Q0 = e * Q0 / norm(Q0)
 end