+++ src/dimreduction/dyca.jl
@@ -2,37 +2,14 @@ export dyca
 using LinearAlgebra
 
 """
-    matrix_gradient(matrix)
-Compute the gradient of a matrix along axis=1.
-
-## Description
-Compute the gradient using second order accurate central differences in the interior points
-and first order accurate one-sides differences at the boundaries. We find the standard
-second order approximation by using:
-```math
-\\hat{f}_i^{(1)} = \\frac{f(x_{i+1}-f(x_{i-1})}{2h} + O(h^2)
-```
-The returned gradient matrix hence has the same shape as the input array. Here we compute
-the gradient along axis=1 (row-wise), so to compute gradient along axis=2 (column-wise),
-the tranpose of the input matrix must be given.
-
-[^Quarteroni2007]: Quarteroni A., Sacco R., Saleri F. (2007) Numerical Mathematics (Texts in Applied Mathematics). New York: Springer.
-"""
-function matrix_gradient(matrix::Matrix)
-    gradient = copy(matrix)
-    gradient[1,:] = (matrix[2,:] .- matrix[1,:])
-    gradient[end,:] = (matrix[end,:] .- matrix[end-1,:])
-    gradient[2:end-1,:] = (matrix[3:end,:] .- matrix[1:end-2,:]) .*0.5
-    return gradient
-end
-
-"""
     dyca(data, eig_thresold) -> eigenvalues, proj_mat, projected_data
 Compute the Dynamical Component analysis (DyCA) of the given `data` [^Uhl2018]
 used for dimensionality reduction.
 
 Return the eigenvalues, projection matrix, and reduced-dimension data
 (which are just `data*proj_mat`).
+## Keyword Arguments
+* norm_eigenvectors=false : if true, normalize the eigenvectors
 
 ## Description
 Dynamical Component Analysis (DyCA) is a method to detect projection vectors to reduce
@@ -63,8 +40,7 @@ is projected onto.
 [^Uhl2018]: B Seifert, K Korn, S Hartmann, C Uhl, *Dynamical Component Analysis (DYCA): Dimensionality Reduction for High-Dimensional Deterministic Time-Series*, 10.1109/mlsp.2018.8517024, 2018 IEEE 28th International Workshop on Machine Learning for Signal Processing (MLSP)
 """
 dyca(A::Dataset, e) = dyca(Matrix(A), e)
-function dyca(data, eig_thresold::AbstractFloat)
-
+function dyca(data, eig_thresold::AbstractFloat; norm_eigenvectors::Bool=false)
     derivative_data = matrix_gradient(data)  #get the derivative of the data
     time_length = size(data,1) #for time averaging
 
@@ -77,6 +53,7 @@ function dyca(data, eig_thresold::AbstractFloat)
 
     #solve the generalized eigenproblem
     eigenvalues, eigenvectors = eigen(C1*inv(C0)*transpose(C1),C2)
+    norm_eigenvectors && normalize_eigenvectors!(eigenvectors) 
     eigenvectors = eigenvectors[:, vec(eig_thresold .< abs.(eigenvalues) .<= 1.0)]
     if size(eigenvectors,2) > 0
         mul!(C3, inv(C1), C2)
@@ -85,4 +62,36 @@ function dyca(data, eig_thresold::AbstractFloat)
         throw(DomainError("No generalized eigenvalue fulfills threshold!"))
     end
      return eigenvalues, proj_mat, data*proj_mat
+end
+
+function normalize_eigenvectors!(eigenvectors)
+    for i=1:size(eigenvectors,2)
+        eigenvectors[:,i] = normalize(eigenvectors[:,i])
+    end
+end
+
+
+"""
+    matrix_gradient(matrix)
+Compute the gradient of a matrix along axis=1.
+
+## Description
+Compute the gradient using second order accurate central differences in the interior points
+and first order accurate one-sides differences at the boundaries. We find the standard
+second order approximation by using:
+```math
+\\hat{f}_i^{(1)} = \\frac{f(x_{i+1}-f(x_{i-1})}{2h} + O(h^2)
+```
+The returned gradient matrix hence has the same shape as the input array. Here we compute
+the gradient along axis=1 (row-wise), so to compute gradient along axis=2 (column-wise),
+the tranpose of the input matrix must be given.
+
+[^Quarteroni2007]: Quarteroni A., Sacco R., Saleri F. (2007) Numerical Mathematics (Texts in Applied Mathematics). New York: Springer.
+"""
+function matrix_gradient(matrix::Matrix)
+    gradient = copy(matrix)
+    gradient[1,:] = (matrix[2,:] .- matrix[1,:])
+    gradient[end,:] = (matrix[end,:] .- matrix[end-1,:])
+    gradient[2:end-1,:] = (matrix[3:end,:] .- matrix[1:end-2,:]) .*0.5
+    return gradient
 end